Move interrupt and exception state files to the separate directory.
