+++ os/kernel/src/cpu/cpuid/cpuid.c
