+++ os/kernel/src/cpu/idt/idt.h
@@ -8,11 +8,11 @@
 #include <string.h>
 #include "idt_entry.h"
 #include "idt_info.h"
-#include "interrupt_state.h"
-#include "exception_state.h"
-#include "interrupt_handler_definition.h"
-#include "exception_handler_definition.h"
-#include "exception_definition.h"
+#include "state/interrupt_state.h"
+#include "state/exception_state.h"
+#include "interrupts/interrupt_handler_definition.h"
+#include "exceptions/exception_handler_definition.h"
+#include "exceptions/exception_definition.h"
 #include "cpu/pic/pic.h"
 #include "cpu/panic/panic_screen.h"
 #include "drivers/keyboard/keyboard.h"
