+++ os/kernel/src/process/manager/process_info.h
@@ -2,7 +2,7 @@
 #define PROCESS_INFO_H
 
 #include <stdbool.h>
-#include "cpu/idt/interrupt_state.h"
+#include "cpu/idt/state/interrupt_state.h"
 
 typedef enum process_status
 {
