+++ os/kernel/src/process/syscalls/handlers/heap_calls.h
@@ -1,7 +1,7 @@
 #ifndef HEAP_CALLS_H
 #define HEAP_CALLS_H
 
-#include "cpu/idt/interrupt_state.h"
+#include "cpu/idt/state/interrupt_state.h"
 #include "memory/heap/heap.h"
 
 void syscall_heap_alloc_memory(interrupt_state *state);
