+++ os/kernel/src/process/syscalls/handlers/pc_speaker_calls.h
@@ -1,7 +1,7 @@
 #ifndef PC_SPEAKER_CALLS_H
 #define PC_SPEAKER_CALLS_H
 
-#include "cpu/idt/interrupt_state.h"
+#include "cpu/idt/state/interrupt_state.h"
 #include "Drivers/PCSpeaker/pc_speaker.h"
 
 void syscall_pc_speaker_enable_sound(interrupt_state *state);
