+++ os/kernel/src/process/syscalls/handlers/rtc_calls.h
@@ -1,7 +1,7 @@
 #ifndef RTC_CALLS_H
 #define RTC_CALLS_H
 
-#include "cpu/idt/interrupt_state.h"
+#include "cpu/idt/state/interrupt_state.h"
 #include "Drivers/RTC/rtc.h"
 
 void syscall_rtc_get_time(interrupt_state *state);
